+++ services/madoc-ts/src/frontend/shared/hooks/use-user-details.ts
@@ -1,18 +1,10 @@
-import { useApi } from './use-api';
-import { useQuery } from 'react-query';
+import { apiHooks } from './use-api-query';
 import { useUser } from './use-site';
 
 export function useUserDetails() {
-  const api = useApi();
   const user = useUser();
 
-  const { data } = useQuery(
-    'user-details',
-    () => {
-      return api.getUserDetails();
-    },
-    { retry: false, enabled: !!user }
-  );
+  const { data } = apiHooks.getUserDetails(() => (user ? [] : undefined), { retry: false });
 
   return data;
 }