+++ CHANGELOG.md
@@ -20,6 +20,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added new experimental [Universal Viewer](https://github.com/universalViewer/universalviewer) page block
 - [Internal] Added new "api.crowdsourcing" namespace to consolidate crowdsourcing APIs together.
 - [internal] Added better debugging names for capture model slots in React devtools
+- Added more APIs to `@madoc.io/types` package (still experimental) for plugins to use.
 
 ### Fixed
 
@@ -46,6 +47,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Fixed keyboard navigation in top right menu
 - Fixed broken manifest link in activity stream and exported manifests
 - [internal] Fixed corrupted translation saving in development
+- Fixed server-side rendering bug with dropdowns
 
 ### Removed